miro_gpt/requirements.txt
absl-py==2.1.0
actionlib==1.13.2
aiofiles==22.1.0
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
angles==1.9.13
anyio==3.7.1
apriltag==0.0.16
apturl==0.5.2
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==19.3.0
audioread==3.0.1
autobahn==17.10.1
Automat==0.8.0
backcall==0.2.0
base-local-planner==1.17.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.7.0
bondpy==1.8.6
breezy==3.0.2
Brlapi==0.7.0
cachetools==5.3.3
camera-calibration==1.16.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==0.4.24
catkin-pkg-modules==0.4.24
catkin-tools==0.8.5
cbor==1.0.0
certifi==2019.11.28
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.3
command-not-found==0.3
config==0.5.1
configobj==5.0.6
constantly==15.1.0
controller-manager==0.19.6
controller-manager-msgs==0.19.6
controller-manager-tests==0.19.6
cryptography==2.8
cupshelpers==1.0
cv-bridge==1.16.0
cycler==0.10.0
Cython==0.29.14
dash==2.16.1
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclass-factory==2.16
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.7
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.16
drone-wrapper==1.4.2
dulwich==0.19.15
duplicity==0.8.12.0
dynamic-reconfigure==1.7.2
dynamixel-sdk==3.7.51
elevenlabs==1.0.1
empy==3.3.2
entrypoints==0.3
exceptiongroup==1.2.0
executing==2.0.1
fasteners==0.14.1
fastimport==0.9.8
flask==3.0.2
flatbuffers==24.3.25
frozenlist==1.4.1
future==0.18.2
gast==0.4.0
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gazebo-video-monitor-utils==0.7.0
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.15
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpg===1.13.1-unknown
grpcio==1.62.1
gTTS==2.5.1
h11==0.14.0
h5py==3.10.0
html5lib==1.0.1
httpcore==1.0.4
httplib2==0.14.0
httpx==0.27.0
hyperlink==19.0.0
idna==2.8
image-geometry==1.16.0
importlib-metadata==7.1.0
imutils==0.5.4
incremental==16.10.1
iniconfig==2.0.0
interactive-markers==1.12.0
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.1.0
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
keras==2.13.1
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
laser-geometry==1.6.7
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-loader==0.3
libclang==18.1.1
libopencv==0.0.1
librosa==0.10.1
llvmlite==0.41.1
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.1.2
matplotlib-inline==0.1.6
mavros==1.13.0
message-filters==1.15.14
monotonic==1.5
more-itertools==4.2.0
moveit-commander==1.1.10
moveit-core==1.1.10
moveit-python==0.4.3
moveit-ros-planning-interface==1.1.10
moveit-ros-visualization==1.1.10
mpi4py==3.0.3
msgpack==1.0.8
multidict==6.0.5
nest-asyncio==1.6.0
netifaces==0.10.4
nose==1.3.7
numba==0.58.1
numpy==1.24.3
oauthlib==3.1.0
olefile==0.46
openai==0.28.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
osrf-pycommon==2.0.1
packaging==24.0
pandas==2.0.3
paramiko==2.6.0
parso==0.8.3
pbkdf2==1.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==4.2.0
plotly==5.20.0
pluggy==1.4.0
pooch==1.8.1
pr2-power-board==1.1.10
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.5.1
pure-eval==0.2.2
pvcobra==2.0.2
pvporcupine==3.0.2
py-trees==0.7.6
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pyassimp==4.1.4
pycairo==1.16.2
pycparser==2.21
pycryptodomex==3.6.1
pycups==1.9.73
pydantic==1.10.14
pydot==1.4.1
pydub==0.25.1
PyGithub==1.43.7
pygments==2.17.2
PyGObject==3.36.0
pygraphviz==1.5
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyRFC3339==1.1
pyserial==3.4
pytest==7.4.4
pytest-asyncio==0.20.3
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.9.0.post0
python-debian===0.1.36ubuntu1
python-gitlab==2.0.1
python-gnupg==0.4.5
python-qt-binding==0.4.4
python-snappy==0.5.3
PyTrie==0.2
pytz==2024.1
pyxdg==0.26
PyYAML==5.3.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
reportlab==3.5.34
requests==2.31.0
requests-oauthlib==2.0.0
requests-unixsocket==0.2.0
resource-retriever==1.12.7
retrying==1.3.4
roman==2.0.0
rosbag==1.15.14
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosdep==0.21.0
rosdep-modules==0.21.0
rosdistro==0.8.3
rosdistro-modules==0.8.3
rosgraph==1.15.14
rosinstall==0.7.8
rosinstall-generator==0.1.22
roslaunch==1.15.14
roslib==1.15.8
roslint==0.12.0
roslz4==1.15.14
rosmake==1.15.8
rosmaster==1.15.14
rosmsg==1.15.14
rosnode==1.15.14
rosparam==1.15.14
rospkg==1.4.0
rospkg-modules==1.4.0
rospy==1.15.14
rosserial-arduino==0.9.2
rosserial-client==0.9.2
rosserial-python==0.9.2
rosservice==1.15.14
rostest==1.15.14
rostopic==1.15.14
rosunit==1.15.8
roswtf==1.15.14
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.11
rqt-controller-manager==0.19.5
rqt-dep==0.4.12
rqt-drone-teleop==1.4.2
rqt-ez-publisher==0.6.1
rqt-graph==0.4.14
rqt-ground-robot-teleop==1.4.2
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-image-view==0.4.16
rqt-joint-trajectory-controller==0.19.0
rqt-launch==0.4.9
rqt-logger-level==0.4.11
rqt-moveit==0.5.10
rqt-msg==0.4.10
rqt-multiplot==0.0.12
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-pr2-dashboard==0.4.0
rqt-publisher==0.4.10
rqt-py-common==0.5.3
rqt-py-console==0.4.10
rqt-py-trees==0.4.1
rqt-reconfigure==0.5.4
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.14
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.9
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.2
rqt-top==0.4.10
rqt-topic==0.4.13
rqt-web==0.4.10
rsa==4.9
rviz==1.14.14
scikit-learn==1.3.2
scipy==1.10.1
SecretStorage==2.3.1
sensor-msgs==1.13.1
service-identity==18.1.0
shazamio==0.5.1
shazamio-core==1.0.7
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
sniffio==1.3.1
soundfile==0.12.1
soupsieve==1.9.5
soxr==0.3.7
srdfdom==0.6.3
stack-data==0.6.3
systemd-python==234
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==1.1.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
threadpoolctl==3.4.0
tomli==2.0.1
topic-tools==1.15.14
tqdm==4.66.2
traitlets==5.14.2
turtlebot3-example==1.2.5
turtlebot3-teleop==1.2.5
Twisted==18.9.0
txaio==2.10.0
typing-extensions==4.5.0
tzdata==2024.1
u-msgpack-python==2.1
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unique-id==1.0.6
urdfdom-py==0.4.6
urllib3==1.25.8
usb-creator==0.3.7
vcstools==0.1.42
wadllib==1.3.3
wcwidth==0.2.13
webencodings==0.5.1
webrtcvad==2.0.10
websockets==12.0
werkzeug==3.0.2
wifi==0.3.8
wrapt==1.11.2
wsaccel==0.6.2
wstool==0.1.18
xacro==1.14.13
xkit==0.0.0
yarl==1.9.4
zipp==1.0.0
zope.interface==4.7.1

